fix(sessions): satisfy organization checks
