+++ packages/gateway-protocol/src/schema/sessions.ts
@@ -12,6 +12,12 @@ import { SessionOwnerSchema } from "./sessions-row.js";
 
 export { SessionsCreateParamsSchema };
 export { SessionsRecoverParamsSchema, SessionsRecoverResultSchema };
+export {
+  SessionsListParamsSchema,
+  SessionsSearchParamsSchema,
+  type SessionsListParams,
+  type SessionsSearchParams,
+} from "./sessions-list-search.js";
 export { SessionsResolveParamsSchema, type SessionsResolveParams } from "./sessions-resolve.js";
 export {
   SESSIONS_PATCH_MANY_MAX_TARGETS,
@@ -376,73 +382,6 @@ export const SessionsDiffResultSchema = closedObject({
       Type.Literal("unknown_commit"),
     ]),
   ),
-});
-
-/** Lists sessions with optional scope, activity, label, and preview filters. */
-export const SessionsListParamsSchema = closedObject({
-  /** Maximum rows to return; omitted Gateway RPC calls use a bounded default. */
-  limit: Type.Optional(Type.Integer({ minimum: 1 })),
-  offset: Type.Optional(Type.Integer({ minimum: 0 })),
-  activeMinutes: Type.Optional(Type.Integer({ minimum: 1 })),
-  /** Require a real user/channel interaction; excludes synthetic isolated heartbeat rows. */
-  requireLastInteraction: Type.Optional(Type.Boolean()),
-  sortBy: Type.Optional(Type.Union([Type.Literal("updatedAt"), Type.Literal("lastInteractionAt")])),
-  includeGlobal: Type.Optional(Type.Boolean()),
-  includeUnknown: Type.Optional(Type.Boolean()),
-  /** Limit agent-scoped rows to agents currently present in config. */
-  configuredAgentsOnly: Type.Optional(Type.Boolean()),
-  /**
-   * Read a bounded transcript head projection to derive a title from the first user message.
-   * Use `limit` to bound projection work on large stores.
-   */
-  includeDerivedTitles: Type.Optional(Type.Boolean()),
-  /**
-   * Read a bounded transcript tail projection for the latest visible user or assistant text.
-   * The returned short preview excludes tool, system, reasoning, and silent rows.
-   */
-  includeLastMessage: Type.Optional(Type.Boolean()),
-  label: Type.Optional(SessionLabelString),
-  /** Filter by custom category; null selects sessions without one. */
-  category: Type.Optional(Type.Union([SessionLabelString, Type.Null()])),
-  /** Limit rows to sessions with an explicitly stored Control UI face preference. */
-  boardFace: Type.Optional(Type.Union([Type.Literal("chat"), Type.Literal("dashboard")])),
-  /** Filter rows by their immutable creator provenance. */
-  creatorId: Type.Optional(NonEmptyString),
-  /** Filter rows by their current assignable owner identity. */
-  ownerId: Type.Optional(NonEmptyString),
-  unread: Type.Optional(Type.Boolean()),
-  status: Type.Optional(
-    Type.Union([
-      Type.Literal("queued"),
-      Type.Literal("running"),
-      Type.Literal("done"),
-      Type.Literal("failed"),
-      Type.Literal("killed"),
-      Type.Literal("timeout"),
-    ]),
-  ),
-  projectId: Type.Optional(NonEmptyString),
-  hasWorktree: Type.Optional(Type.Boolean()),
-  needsAttention: Type.Optional(Type.Boolean()),
-  /** Limit rows to sessions owned by or previously prompted by the authenticated viewer. */
-  involvingMe: Type.Optional(Type.Boolean()),
-  spawnedBy: Type.Optional(NonEmptyString),
-  agentId: Type.Optional(NonEmptyString),
-  search: Type.Optional(Type.String()),
-  /**
-   * True lists archived sessions; "all" lists archived and active;
-   * false or omitted lists active sessions.
-   */
-  archived: Type.Optional(Type.Union([Type.Boolean(), Type.Literal("all")])),
-});
-
-/** Searches one agent's indexed session transcripts, optionally within selected sessions. */
-export const SessionsSearchParamsSchema = closedObject({
-  agentId: Type.Optional(NonEmptyString),
-  sessionKeys: Type.Optional(Type.Array(NonEmptyString, { minItems: 1, maxItems: 200 })),
-  query: Type.String({ minLength: 1, maxLength: 4096 }),
-  resultMode: Type.Optional(Type.Union([Type.Literal("messages"), Type.Literal("sessions")])),
-  limit: Type.Optional(Type.Integer({ minimum: 1, maximum: 100 })),
 });
 
 /** One full-text session transcript match with follow-up provenance. */
@@ -845,11 +784,9 @@ export const SessionsUsageParamsSchema = closedObject({
 
 // Wire types derive directly from local schema consts so public d.ts graphs never
 // pull in the ProtocolSchemas registry.
-export type SessionsListParams = Static<typeof SessionsListParamsSchema>;
 export type SessionsCleanupParams = Static<typeof SessionsCleanupParamsSchema>;
 export type SessionsPreviewParams = Static<typeof SessionsPreviewParamsSchema>;
 export type SessionsDescribeParams = Static<typeof SessionsDescribeParamsSchema>;
-export type SessionsSearchParams = Static<typeof SessionsSearchParamsSchema>;
 export type SessionsSearchHit = Static<typeof SessionsSearchHitSchema>;
 export type SessionsSearchResult = Static<typeof SessionsSearchResultSchema>;
 export type SessionCompactionCheckpoint = Static<typeof SessionCompactionCheckpointSchema>;
